+++ hbs-core/src/main/java/pl/edu/agh/hbs/core/model/ui/cartesian/ViewPosition.java
@@ -3,30 +3,30 @@ package pl.edu.agh.hbs.core.model.ui.cartesian;
 import java.util.Objects;
 
 public class ViewPosition {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public ViewPosition() {
     }
 
-    public ViewPosition(int x, int y) {
+    public ViewPosition(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
